+++ app/[search-job-offers]/page.tsx
@@ -8,7 +8,7 @@ import SearchBox from "../ui/search-box/SearchBox";
 import PrimaryButton from "../ui/common/buttons/PrimaryButton";
 import ModeWrapper from "../ui/common/ModeWrapper";
 import JobOfferList from "../ui/common/jobOffer/JobOfferList";
-import { findJobOffers } from "../lib/actions";
+import { findJobOffers } from "../lib/actions/filterJobOffersActions";
 import { JobOfferHeader } from "@/models/JobOfferHeader";
 
 const SearchedJobOffers = async () => {
